Create html-product-selector
Provides a way of using the Shopify Buy Button with a custom drop-down selector for size (or other product variants). Currently only created to work with the buy now method of purchasing and with a single quantity. Future updates will add second form for quantity and possibility to add to cart instead of purchase. 

In order to use, replace the long numbers with the unique product ID. It can be found here: https://help.shopify.com/themes/customization/products/find-variant-id